Ta bort kort *work in progress
*Har justerat add-card-funktionen så att den lägger till en ta-bort-knapp
i varje card.
*Har delat upp koden så att funktionerna addcard (och snart deleteCard)
ligger i page-modulen.
*Har anpassat CSS-koden i Card för att kunna flytta min delete-card-btn
till övre höger hörna ====> .card har nu position: absolute. Jag kan
inte se några förändringar/buggar som detta kan ha medfört.

/Anthon
